kitt: Android: Fix the swipe to delete bug
When you swipe to delete multiple items in one go, the sequence of
functions that are called to perform this task get interrupted halfway.
This leads to the first item that was swiped not being deleted in the
end.

To fix this, add a check before the beginning of every such sequence to
make sure another sequence is not happening. If so, do not wait for a
dismissal of the snackbar and force delete the item from the db.

Signed-off-by: Nikhil Punathil <nikhilpe@gmail.com>